Define and run two separate test suites
The idea is to enable faster local development of the tests by
seperating `gnome_install` (which takes 2+ minutes to complete in
all cases) from subsequent tests such as `gnome_apps` which can complete
within seconds.

There are two GNOME OS artifacts - the ISO installer and the
preinstalled HDD image. The `gnome_apps` tests now use this second
artifact to skip the install step.

In Gitlab CI everything still runs sequentally so there is no speed
boost to the CI. (In fact it's a little slower as it needs to fetch
and unpack the disk image). We absolutely can split it into separate,
parallel CI jobs when needed.

Multiple test suites are defined in `main.pm` by reading the `TEST`
config variable which is set based on the testsuite name.

Note: I did also try saving the disk image from the gnome_install step
and using that in the gnome_apps test. OpenQA requires the image to be
uploaded to the web UI machine then dowloaded again between tests and
this added a 10+ minute slowdown to the tests.

https://gitlab.gnome.org/GNOME/openqa-tests/-/issues/3