Move scripts from utils/ into the root directory
The openqa-utils.git repo was setup by Martín for use by the module QA
tests. We need to have this stuff only in one place, and the simplest
way seems to be to replace the current openqa-tests.git/utils directory
with openqa-utils.git using a Git submodule.

This will probably break the GNOME Shell QA prototype at
<https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/3419>
but that's why it's still a prototype ;)

Part of: https://gitlab.gnome.org/GNOME/openqa-tests/-/issues/127